Fixed a couple build errors

+++ moka/io/buffer.h
@@ -42,7 +42,7 @@ class Buffer;
 
 class moka::io::Buffer {
 public:
-  static v8::Handle<v8::Value> New(uint32_t size = 0);
+  static v8::Handle<v8::Value> New(size_t size = 0);
 
   static v8::Handle<v8::FunctionTemplate> GetTemplate();
